Extend spout classes with a getXmlUrl function
This adds a new function to the spout interface and its subclasses:
spout::getXmlUrl(array $params)

getXmlUrl receives the same $params array as the spout::load function.

getXmlUrl can use the $params argument to determine the URL of the RSS
feed used by the source and should return a single URL as a string.
If the spout does not fetch its content from an RSS feed internally,
(e.g. because it uses an API instead), getXmlUrl should return false.