Middle click on the title to open the entry in an new window
Some notes about the implementation:
Actually JavaScript neither opens the window nor follows the link. The user clicks on an regular <a href="">. We just prevent the browser from following the link on a regular left click. This is done with preventDefault() within an click event.
Because Firefox never triggers the click event on middle click, we must use the mousedown and mouseup events. Unfortunately Firefox even doesn't trigger the mouseup event by default. To work around that, we must call preventDefault() within the mousedown event. Now we can mark the entry read within the mouseup event. Finally, we use .data() to prevent fragmentary clicks.