Fix pgsql VACUUM ANALYZE syntax error
Unfortunately, using a bound parameter for a table name (as opposed to a value) is not supported by PostgreSQL, so "VACUUM ANALYZE :table" is a syntax error.  We have to dynamically create this query.  See this thread for details:

http://www.postgresql.org/message-id/CAFj8pRD1Y4tpAzXeQR511+q8qJ9a5n4CpF=64HrudQtWRraZiw@mail.gmail.com

Since the parameter in this query is a table name in the database, any attacker would have to have create table privileges, so I don't consider this a SQL injection risk.