do not format long articles in columns
Because in order to read them when they are longer than the vertical view pane,
you need to scroll up and down and again up and down and one last time up
and down to read the three columns. It is preferable in that case not to
format in columns.