Add session keep alive
Send periodic AJAX requests to keep the session alive. About every 10 minutes a request is sent, which actually does nothing, but prevents PHP from garbage collecting the session file. When the user is neither logged in nor Selfoss is in public mode, the users session expired and can't be re-initialized automatically, as it would be possible in public mode. If this is the case, reload the page, which will present the login form.