Add functionality to show when the last post was found for a source.
In the "(add) sources" list in the web browser the number of days since a new post was
found for each source will be shown to the user.

The functionality is implemented by adding a new field to the sources table into which
the date/time is written when a new post was found for the source.