daos: Fix getting a single source
1973cc954cbd44c37d5583a8b3f9bc822a805930 introduced an optional argument
to `\daos\mysql\Sources::get` for getting a single source. The commiter,
however, forgot to also add it to `\daos\Sources::get`, which means that
the commit never worked. The MySQL method never received any arguments,
resulting in the password being deleted, instead of preserved, when user
did not indent to change it.