ensure entry date can be used by the client side js code
...and use ISO8601.

This triggered an inconsistency for pgsql wich was recording
datetimes with millionth of a second precision whereas ISO8601 only
handles thousandth of a second precision. This would have caused problems
of duplicate entries for stream more and new offset_from_datetime
implementation.
That's why pgsql datetimes columns are converted to second precision which
is enough.