prevent ignoring new items from other sources with the same uid
selfoss was checking for items with the same uid in the whole items database,
not within the same source. This change searchs for items with the same uid
only in the same source.

This would have potentially allowed one feed to prevent any new items from
being loaded from another feed.