make it unlikely that the client browser gets outdated js or css
This change makes the js and css generated to a file containing the
max mtime of source material in its name. Therefore, each time the
source js or css files are changed, a new file with a new name gets
generated and referenced in the app. From the client browser's view,
this is a new ressource that needs to get fetched regardless of the
cache status of older css or js.

The trade-off is that each request implies stat'ing each source
material file. There is no notiecable time penalty in loading time.