Revert "Prefer item link for obtaining favicon"
When RSS feed item link redirects to a different site (e.g. Feedburner)
and there are no icon link elements, favicon.ico in the root of the site
will be checked. Unfortunately, the site root is determined from the original
item link before redirection, therefore incorrect icon might be used. Until
we can determine the correct site root easily, possibly once we switch
to Guzzle, the original behaviour is preferred.

Additionally, the commit caused that the links in link aggregators
like Haskell News used the favicon of the first link.

This reverts commit cac783d75cd6d81ac545879d5b45f39432236d20.