Do not use cache unnecessarily
When loading an image, the data were written to a file to determine
the file type. This file was previously cleaned but the clean-up was
accidentally dropped in d1307573d3d75a7a8078c8e2942c44ba6aa63036.

PHP 5.4 introduced `getimagesizefromstring` function, which allows
to use data already in memory, removing the need for writing
to a temporary file and as a result a clean-up.