Make source code checkers accessible from Grunt
To make the development cycle faster we want developers to be able to run
all the required checks locally. Because of the way Composer is designed,
we cannot save the checkers as dev dependencies but have to install them
separately.

Instead of using composer for installing linters to utils directory, we
now use the tools from PATH, allowing users to install them using their
package manager of choice. For Travis and users running distributions
that do not have the checkers packaged, there is `utils/install-phars.sh`
helper script, which will install the necessary checkers to ~/.local/bin
directory.

This commit introduces new grunt tasks:

* `lint:server` will check PHP syntax using built-in PHP tool
* `cs:server` will check the coding style using PHP-CS-Fixer
* `check:server` will run both the checkers (more in the future)
* `check` will run both server side and client side checkers