Fix fancybox being considered extraneous package
Since Debian-provided version of NPM does not support scoped
packages, we tried to install FancyBox from Git. I also changed the
dependency name in the `package.json` to just `fancybox` in order
not to trigger the outdated NPM. Apparently, this made NPM
think the package is extraneous due to a mismatch between the
package name and the name in the `package.json`. And what is more
it does not work with older NPM versions either.

For that reason, we have decided to ignore the older NPM versions
for now.