Fix an error when migrating items with updatetime=null
In 14442c03b882a8f4e2255c5b654459a7bd8c5f78, `updatetime` field was
introduced to the `items` table. Unfortunately, unlike the
initialization, the migration did not flag the field `NOT NULL`. This
resulted in the pre-existing items to have `updatetime = 0`. When
re-creating the table for adding an another field in
83794f415e4ba8b8339b1236155d52769e799dc6, the `updatetime` was set to
`NOT NULL`, based on the previous initialization. That caused an error
when migrating those items.

This patch fixes the assumption by filling the `NULL` fields.

Closes: #954