Fix navigation on mobile
In 8946e49e75e8b82815e341f683badf2a40e97e79, we upgraded the scrollbar
plug-in. Due to [changes] in version 3.1.0 of the plugin, the height of
a hidden container is set to zero. This caused the navigation to be
invisible in mobile layout.

With this patch, the scrollbar library will only be initialized when
using desktop layout.

Closes: #953

[changes]: https://github.com/malihu/malihu-custom-scrollbar-plugin/commit/12513564a32dbb7638dbf70b3d9d8b3f1b6655e8#diff-379285ae10c1a5aad9c2ba1ccd07dd96L517