Use PSR-4 for autoloading
When we introduced the use of Composer classmaps for autoloading in
df6f5d033ff73c4a29e9118abcb0ae7968710009, we lost the ability to define
custom spouts without dumping the autoloader. PSR-4 would fix this issue
but at the time I came to conclusion we would have to change selfoss
file structure to support it.

Since that assumption seems to be false, we are switching to PSR-4, thus
having the cake and eating it too. The only caveat is the need for more
strict file naming conventions.

Closes: #958

### Upgrading

If you use any custom classes like spouts, make sure they follow the
[PSR-4], namely the namespace should match the file path (even in letter
case).

For example, if you have custom spout class named `PrivateSpout` in
`spouts\private` namespace, it will have to be located in
`spouts/private/PrivateSpout.php`.

This also means each file can only contain a single class.

[PSR-4]: http://www.php-fig.org/psr/psr-4/

### Upgrading (developers)
Once you checked for compliance, run `composer install` as described in
the [development section] of README to switch to PSR-4 autoloader.

[development section]: https://github.com/SSilence/selfoss/blob/master/README.md#development