client: Log out user instead of showing 403 Forbidden error
Pull Request #931 moved the authentication client side, which worsened
the issue #919 by displaying error instead of the login form on opening
selfoss when user did not log out and their session expired on the
server.

This patch redirects user to login form whenever client side session is
registered and 403 Forbidden error is received.

This is a temporary fix before offline support is merged.