Fix compatibility with PaaSs
On Heroku and other services, client side resources are not installed
since the client package.json is in a subdirectory. We need to install
them manually.

This commit adds the client installation script into postinstall hook.

Please note that you still need to enable node.js buildpack (or carry
out an equivalent procedure for your PaaS) for the hook to be run.