client: Change icons to vector
Vector icons will look much crisper on HiDPI devices. Using them will also allow us to colour them in a dark theme.

We use FontAwesome DOM loading script for now. We will be able to use the `icon` helper directly in JSX when we complete the switch to client side rendering.

For size reasons, we are subsetting the icons, so when we want to use an icon not already used, we will need to add it here. But this was true for image files as well.

We need to use deep imports for plug-ins, though, since Parcel’s tree shaking is shaky: https://github.com/SSilence/selfoss/pull/1013#issuecomment-538638759

We also no longer depend on parcel-plugin-static-files-copy, since we only used it for icons referrence by <img> tag from PHP templates (not picked up by Parcel).