spouts/reddit2: remove special imgur handling
The Reddit spout tried to detect if an item points to imgur.com and
change the link directly to an image in that case. Unfortunately,
it was flaky, for example, gifv files resulted in 403 error.

Instead we are using the previews feature, which, as a bonus, also
works on sites other than imgur.