ui: Stop bundling Open Sans font
The font is quite big but it still misses many glyphs. It is also not updated
regularly. Leaving it out will make the distribution much smaller and selfoss
will load faster.

If users still prefer it, they can install it to their devices. Or just copy
the removed fonts.css to custom.css.