helpers/image: Improve icon selection algorithm
Previously, we chose the icon from the HTML based on their type and when the link was broken we did not try another one. Additionally, we previously [1] accidentally stopped recognizing icons that used uppercase letters.

This commit fixes both by switching to a more robust regex-based parser and returning array of icons sorted by their size, that will be checked until we find a working one.

[1]: https://github.com/SSilence/selfoss/commit/6e0494c229803674499c634a3e3b2a4906add78f

Fixes #1122

Co-Authored-By: Jan Tojnar <jtojnar@gmail.com>