a11y: Make sidebar links more accessible
The headings for "filters", "tags" and "sources" are clickable like a link so, added the link role as well along with the heading.
There was no need of aria-haspopup attribute, so removed all the references.
Also, Filters, Tags etc are defined as a list but role="link" added to the list item because they work like a link, thus, screen readers do not treat them as a list, thus, added the role="listitem" also in the span so that screen reader properly recognise them as both list and link.

Co-Authored-By: Jan Tojnar <jtojnar@gmail.com>