client: refactor aria-expanded manipulation to function calls
Also wrap the contents of `.entry-title` with `span[role=link].entry-title-link` which is now the holder for the `aria-expanded` attribute. This will improve the accessibility of the entry page. Unfortunately, it might cause some issues with tweets, that can contain links in the title.

Also change the focus target when navigating using keyboard to that link, since it is tabindexed, unlike the icon, used previously.