api: Add /api/about endpoint
This will allow clients to obtain version of Selfoss and API, as well as the the instance configuration. The configuration options are not part of the public API yet, though.

I put it under /api prefix, as we might move the rest of the API there in the future.

Increases API version to 2.20.0 – new feature