dist: Build in a separate repo
Building dist archive from the current repo can introduce uncommitted impurities and makes it harder to check if the assets are packed correctly. Inspired by Meson, we will clone the repo into a temporary directory and build the archive there.

We also no longer run postinstall scripts from the dist script to speed it up a little. We will need to install the dependencies in the temp dir anyway.