spouts\rss: Provide unencoded link
SimplePie returns links escaped with htmlspecialchars but we might want to use the URL to fetch additional data so we need to unescape it first.
ContentLoader will escape it again before saving it to database.

I went through all the spouts looking for returned values that are used as URLs, Reddit does it too.

Ideally, everything would return raw data and we would escape it when displayed but that would be a pain since we would likely have to convert the already stored data.