client: Fix sources in sidebar stopping being clickable
In 396a3f076a300902d9b4fb98a1e4680a28b8260a, we added an extra `delayNavigation` argument to `refreshSources`. Unfortunately, `reloadList` function called `refreshSources` with an extra argument that was never used and with the change it was passed as `delayNavigation`. This meant the navigation never happened after the item list was reloaded.