client: Fix loss of selected entry on loading more items
In d4e362dae69e11e2f68b1a2add26ce55eafe171e, I refactored the how selected entry is handled but started clearing it too eagerly. This lead to the entry being cleared when more entries were loaded, not only when changing section and dropping the old entries. It was especially annoying when navigating by keyboard because the cursor jumped back to the first entry instead of continuing to the newly loaded items.

Now we will only clear it when we would clear the entry list.