common: Fix Dice injection
Since the dependency injection refactoring, SpoutLoader depends on Dice instance to create
spouts but it was getting a fresh instace instead the global, already configured one. When this
instance was asked to create a Logger, it was not aware of our DIC rules so it created a fresh
Logger instance not respecting the selfoss configuration.

Once we switch to Dice 4, we will be able to simplify the rule a bit:
https://github.com/Level-2/Dice/pull/175

Fixes: https://github.com/SSilence/selfoss/issues/1196