client: add {noscript,script loading,script error} messages
When JavaScript is disabled, selfoss will just show a blank page which is suboptimal.
And even with JavaScript enabled, the page is still blank while it is loading or if the script fails to download.

Let's add a basic messages to let user know what is up.

cc https://github.com/SSilence/selfoss/issues/1199