Catch throwables in ContentLoader
In PHP 8, many functions started to throw an `Error` instead of triggering
`E_ERROR`, `E_WARNING` or other error level as before:

https://wiki.php.net/rfc/consistent_type_errors

This is great in general, but it means the errors ignored by @ operator are
no longer caught and since `Error`s extend `Throwable` instead of `Exception`,
our “sandbox” no longer catches them.

Ideally, we would use an error handler that throws an `Error`, wrapping
the errors on older PHP versions, and replace all @ operators by specific
try-catch blocks but that is a lot of work (also needs to be done in libraries).