Prevent double-escaped entities in the navbar
For legacy reasons, we escape tags and source titles on the server.

Previously, we used server side templates where we did not escape the values but with a switch to client-side templates in https://github.com/fossar/selfoss/pull/1184, we started escaping those texts in the templates. This resulted in double escaped ampersands being shown to users.

Let’s unescape those values client-side for now since unescaping them in the database or on the server would break backwards compatibility for third party clients.

Fixes: https://github.com/fossar/selfoss/issues/1241