helpers: Do not crash on PHP 8 when unable to write to storage
Because of a PHP bug https://bugs.php.net/bug.php?id=80700,
incorrect error_reporting level is set on PHP 8,
which makes the warning crash the loader.

But we should have suppressed the warnings anyway,
since we handle the errors by checking the return values.

And changing the value of error_reporting should be avoided as well
as a matter of good code practice.