client: Upgrade to Parcel 2
Parcel 1 is no longer maintained so let’s switch to beta version of Parcel 2. It is much cleaner and we will be able to use things like HMR.

(We actually need to use nightly because parcel-config-precache-manifest targets that. And HMR is disabled for now due to <https://github.com/parcel-bundler/parcel/issues/6129>.)

Also needed to move the service worker to root of assets/ because Parcel preserves the directory structure for entry points but not for non-entry bundles, which would mean all the assets (ending up in dist/) would be out of scope for service worker in dist/js/. See <https://github.com/parcel-bundler/parcel/issues/6173>