ci: Try to keep latest uploaded package to Cloudsmith marked as such
Cloudsmith currently marks with the “latest” tag the package which has the greatest version according to semver ordering, not the latest uploaded package. Since our nightly builds have commit hash as a part of the version, they get misordered alphabetically.

To prevent this, let’s set the “latest” tag manually.