client: Fix nav links with search active
With the switch to using `Link`s for nav items, we started using location objects,
which have query string as a separate field. But we continued to append
the query string to the pathname field, resulting in a duplicated query string.

Let’s fix that by creating a separate function for creating location-based `To`s
and change the old `makeEntriesLink` for creating string-based `To`s to use it.

While at it, I also made the function preserve the original query string value,
so that only the `search` parameter is modified.