client: “New source” form should not wait on fetching spouts
They are already fetched from the initial loading of sources
so we should not need to wait for them to be fetched again.
In the rare case that they did indeed change, we can update
the select box later.

This will significantly speed up opening the “New source” form,
especially when using the `/manage/sources/add` route.