spouts\youtube: add support for fallback URLs
Unlike with the old YouTube channel URLs, it is not possible
to derive the RSS feed URL from the recent `/c/${channelName}` URLs.

Fortunately, YouTube still includes `link[rel=alternate]` in the HTML
so let’s just pass the URL to SimplePie as is and let it discover the feed.

Fixes: https://github.com/fossar/selfoss/issues/1273