client: Use consistent focus ring
Modern browsers use a thick focus ring in a contrasting colour
(~system/browser-theme specific in Firefox, white for Chromium)
while older ones used thin dotted outline.

Previously, we were styling buttons based on the older browsers’ style
with custom colour for increased visibility. But it was still not visible
enough and what is worse, was inconsistent with focus ring on links
in modern browsers.

Let’s use the modern :focus-visible pseudo-class to enforce the same focus
ring style for all elements. Though, we need to adjust it for some
elements to make it visible.

In older browsers, it will fall back to UA style but it is widely supported
nowadays (88.53% according to https://caniuse.com/css-focus-visible).