client: Nuke off-line data on logout
Might be somewhat annoying when the “log out” button is clicked accidentally but it is generally expected that on log out all session data will be cleared.

Maybe we can hide the button into a submenu in the future.

Also make the `logout` function async and re-organize it a bit (redirect only after user has been logged out).