client: Work around over-escaped search highlighting in source title
Since we ported the client to React, most of the strings are treated as plain text,
which is the default behaviour to prevent XSS and other security issues.
Search mode injects HTML markers into `title`, `content` and `sourcetitle` fields, though.
And while `title` and `content` are already HTML, so we do not need to do anything there,
`sourcetitle` is treated like a plain text so the marker HTML tags would be displayed verbatim.

Let’s strip the markers out and replace them with actual tags.

Ideally, we would just highlight on the client side to also allow searching in offline mode.
But until then this hack will have to do.