ContentLoader: do not sanitize author field
It is supposed to be a plain text field, even though SimplePie escapes it.
We are already decoding it.

We are not updating existing fields in the database since, even though it is improbable
that a text that looks like escaped HTML would be used as an author, it cannot be fully ruled out.
If you want to do it, you can run the following SQL code:

    UPDATE items SET author = REPLACE(REPLACE(REPLACE(author, '&gt;', '>'), '&lt;', '<'), '&amp;', '&')

Fixes: https://github.com/fossar/selfoss/issues/1317