client: Only allow generating passwords for authenticated sessions
Since `password_hash` function is intentionally computationally intensive, allowing to use it publicly could open the server to DoS attacks.

Let’s only allow it when there is no password set or user is logged in.

The downside is that user will not be able to use the page when they forget their password without temporarily disabling the authentication in config.