PHPStan: Raise level to 8
**BC break**: The `$extraData` argument of `Item::__construct` is no longer optional.

We are not suppressing the possible `preg_replace` error since the global error handler will catch it. Ideally, we would throw it explicitly when it returns `null` but the error messages from `preg_last_error_msg()` appear to be inferior to the suppressed error so we would need to juggle error handlers. And even before the error is triggered here, in case of invalid regex, it would fail in the search first.