client: Fix articles containing HTML-special charaters
SimplePie returns links with HTML-special characters escaped.
Before the switch to React, the article URL would be printed by the template as is.
React escapes everything by default, resulting in the link targets being double escaped.
Let’s unescape them on client side for now.
In the future, we will want to do it when accepting the data from SimplePie.

Fixes: https://github.com/fossar/selfoss/issues/1407