spouts/fulltextrss: Fix infinite loop on error
The Graby spout uses the content from the original item when an error occurs.
But during lazification in e017101c634a5b576e7b410e7a22443495247b87,
we accidentally started to use the not yet resolved content from the thunk,
leading to an infinite loop.

The loop was fixed in the parent commit by returning an empty content
when the thunk is not yet resolved.
Here we just fix the fallback to use original content again.

The issue was revealed on MMOSpy by e06aa61ce27ab2ddc6563fc1a80d44bfafcd4619.
The previous URL handling accidentally discarded the port number
and, as a result, masked that the feed provides a broken URL:

    http://www.mmo-spy.de:443/