daos/mysql: Add colors to tags without one
When user specified a source tag with a same name but a different letter casing as another existing tag,
the MySQL backend would think the tag already exists and not create a color for it.
This would break, for example, OPML export, which expects every tag to have a color associated with it.

The bug has been fixed in the previous commit but there can still be such erroneously created tags from before.
Let’s go through all tags mentioned in the `source` table and ensure the `tags` table is up to date.
If a tag with a different casing and a color exists, let’s consider it the canonical casing and update the source tag to that.

Fixes: https://github.com/fossar/selfoss/issues/1438