client: Restore “Mark all” button to 2.18 behaviour
In 2.18, when in “unread” mode, if there were no unread items,
clicking the button would clear out all items and trigger a reload
or refresh status of the “Load more” button, depending on whether
the client was aware of more unread items on the server.

When we ported entries to React in f3b279c808e8a5d7e3ab5c8254bdda83c094b50e,
we accidentally broke the “Mark all as read” button,
making it keep all the items it marked in the view (a).
We also did not port the code that would load more items afterwards (b).

It got further broken during porting routing to React in 405a3ec2cf49e3dc17421b997930c5b1faebd2fc.
We looked for the `filter` field on a wrong object (c).

We fixed (c) and attempted to fix (a) in ec544b7ae066050dcdb43e362f38854d4fdb7bea
but the patch just made it keep all the items it did not mark, such as those marked as read
manually beforehand, instead.

This patch makes it so that the button removes all the items in the view except for the active item
and loads more items afterwards, returning to behaviour similar to 2.18.

Since we currently do not have a sane dispatch method in EntriesPage,
we unfortunately have to trigger a virtual click on the “Load more” button.
To make that work, we also need to keep the old offsets since when we clear the entries,
there would be no way to get the offsets back in the `moreOnClick` handler.