wip: client: Relegate source filter to advanced settings
Most people will not use filter for most sources so let’s only show it
when filled in or when advanced settings are enabled.

With advanced settings we will be able to implement less common features
like archival, reduced update frequency or icon overriding.

WIP because I do not like the the toggle button placement.