WebClient: Send Accept-Encoding
Guzzle does not send `Accept-Encoding` header by default.
That is equivalent to sending `Accept-Encoding: *`:
https://www.rfc-editor.org/rfc/rfc9110#field.accept-encoding>
Most servers will probably return an uncompressed body
in response to that, which can be considered wasteful,
and can trigger crawler detection systems:
https://github.com/fossar/selfoss/issues/1481
Others might even opt to use a compression method
that is not supported by the system
(e.g. when libcurl is not compiled with brotli support).

Let’s force Guzzle to let curl send `Accept-Encoding`
header reflecting which compression methods it supports:
https://github.com/guzzle/guzzle/pull/3215