ci: Use proper SemVer for Cloudsmith
We use `$nextVersion-$commit` as a version string but this confuses Cloudsmith,
which expects version strings to follow Semantic Versioning specification.
As a result, the git commit hash after the dash is considered a pre-release
version so Cloudsmith assigns the automatic `version:latest` tag to
the snapshot with the numerically greatest commit hash (e.g. fffffff).

Let’s use fixed `pre` as the pre-release version, which will be greater
than any commit hash, and relegate the commit hash to build metadata.
The SemVer spec mandates that “Build metadata MUST be ignored when
determining version precedence.”

We might consider switching to this scheme everywhere in the future
but I am skipping that for now since selfoss releases do not really
follow SemVer. Using plus in file names might also still be problematic
for some users – while it is supported by all commonly used systems,
it is a special character for URIs and I do not want to hear about bugs
in third party systems like the following:
https://www.github.com/OneDrive/onedrive-api-docs/issues/555
https://www.github.com/nextcloud/server/issues/15849