client: Use prettier for formatting
I do not like the formatting very much but at least it is consistent and supports TypeScript.
https://typescript-eslint.io/troubleshooting/formatting/

Also replace `lint:js` npm script in the `client/` directory with `check:js:lint`
and add `check:js` which runs that in addition to newly introduced `check:js:prettify`.
This matches how the stylesheet scripts are structured.

Adding `eslint-config-prettier` disables stylistic rules that would fight with Prettier.
Additionally, I had to remove some clashing from our ESLint config.

We are using `eslint-config-prettier` instead of `eslint-plugin-prettier`,
the latter would run prettier as part of ESLint but we are already running it separately.