client: Use prettier for formatting
I do not like the formatting very much but at least it is consistent and supports TypeScript.
https://typescript-eslint.io/troubleshooting/formatting/

Also replaces `lint:js` npm script in the `client/` directory with `check:js:lint`
and adds `check:js` which runs that in addition to newly introduced `check:js:prettify`.
This matches how the stylesheet scripts are structured.

We are using `eslint-config-prettier` instead of `eslint-plugin-prettier`,
the latter would run prettier as part of eslint but we are already running it separately.

TODO: Remove stylistic rules from config.